+++ src/upgrades/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+"""Kubernetes Upgrade Code."""
